Move IStartConfiguration inside DynamoModel.
